requirements.txt
flask
gunicorn
google-generativeai
requests
beautifulsoup4